Move Mix.TextDiff to Mix.Tasks.Format.text_diff_format
